requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Pillow==7.0.0
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.8
